Change back name and instead just add incremental full refresh test to the test_ha_iceberg.py file
